修复删除用户为null时异常报错

+++ src/main/java/com/groupten/online_music/web/UserManageController.java
@@ -13,6 +13,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
+import java.util.NoSuchElementException;
 
 @Api(tags = "用户管理相关接口")
 @RestController
@@ -25,10 +26,10 @@ public class UserManageController {
 
     @ApiOperation(value = "新增用户接口")
     @PostMapping
-    public ResponseEntity add(@RequestParam Map<String, String> userMap){
+    public ResponseEntity add(@RequestParam Map<String, String> userMap) {
         User user = new User(userMap);
         ResponseEntity<User> responseEntity = new ResponseEntity<User>();
-        if(!userManageService.hasUser(user) && userManageService.findByEmail(user.getEmail())==null){
+        if (!userManageService.hasUser(user) && userManageService.findByEmail(user.getEmail()) == null) {
             user = userManageService.save(user);
         } else {
             return responseEntity.message("有重名用户或邮箱已注册");
@@ -40,16 +41,12 @@ public class UserManageController {
     @ApiOperation(value = "用户分页查询接口")
     @GetMapping
     public @ResponseBody
-    ResponseEntity FindAll(@RequestParam Map<String, String> pagingMap){
+    ResponseEntity FindAll(@RequestParam Map<String, String> pagingMap) {
         ResponseEntity<Page<User>> responseEntity = new ResponseEntity<>();
         STablePageRequest tablePageRequest = new STablePageRequest(pagingMap);
         Page<User> page;
-        try {
-            page = Page.empty(tablePageRequest.sTablePageable());
-            page = userManageService.findAll( tablePageRequest.sTablePageable());
-        } catch (ApplicationException ex) {
-            return responseEntity.message("分页查询失败");
-        }
+        page = Page.empty(tablePageRequest.sTablePageable());
+        page = userManageService.findAll(tablePageRequest.sTablePageable());
 
         return responseEntity.message("分页查询成功").data(page);
     }
@@ -60,16 +57,15 @@ public class UserManageController {
     ResponseEntity delete(@PathVariable int id) {
         //响应结果
         ResponseEntity<User> responseEntity = new ResponseEntity<User>();
+        //删除操作
         User target = userManageService.findById(id);
-        EmailConfirm emailConfirm = emailService.findOne(target.getEmail());
-        try {
-            //存在用户，删除相关信息
+        //存在用户，删除相关信息
+        if (target != null) {
             userManageService.delete(target);
-            if (emailConfirm != null)
-                emailService.delete(emailConfirm);
-        } catch (ApplicationException e) {
-            e.printStackTrace();
-            return responseEntity.message("删除请求失败！\n" + e.getCause());
+            EmailConfirm emailConfirm = emailService.findOne(target.getEmail());
+            emailService.delete(emailConfirm);
+        } else {
+            return responseEntity.message("删除请求失败！ 目标对象为null");
         }
 
         return responseEntity.message("删除请求成功！");
